Cleanup RunSelLdrTests commands.
Most don't need to test for ARM/PNaCl, it's now rolled into
WriteSelLdrScript.

BUG=none
R=sbc@chromium.org

Review URL: https://codereview.chromium.org/26444003

git-svn-id: https://naclports.googlecode.com/svn/trunk/src@910 7dad1e8b-422e-d2af-fbf5-8013b78bd812

+++ build_tools/common.sh
@@ -900,6 +900,12 @@ WriteSelLdrScript() {
     # no sel_ldr for arm
     return
   fi
+
+  if [ ! -e ${NACL_IRT} ]; then
+    echo "ERROR: Missing IRT binary. Not running sel_ldr-based tests."
+    exit 1
+  fi
+
   if [ $NACL_GLIBC = "1" ]; then
     cat > $1 <<HERE
 #!/bin/bash
